requirements.txt
anyio==4.12.1
certifi==2026.2.25
charset-normalizer==3.4.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
python-telegram-bot[job-queue]==21.6
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.5
urllib3==2.6.3
